Added crypto processor tests.
