+++ com/src/CryptoProcessor.py
@@ -14,7 +14,7 @@ class CryptoProcessor(Processor):
         self.api_requester = api_requester
         
         #Have one processor for entire engine so this will only be called once in init().
-        self.populate_coin_list()
+        self.populate_coin_hash()
 
     def handle(self, message: BeautifulSoup):
         for message_item in message.findAll(['p','h3']):
@@ -35,17 +35,23 @@ class CryptoProcessor(Processor):
                         currently_seen_coins.append(current_coin)
             self.seen_post_titles.append(post)
         
-
     #Purpose of method is to call api_requester, which will need to be refactored to take url as param.. but anyways,
     #i am storing all coins as a hash table as the look ups are instant, and i want to hash 
     #symbols to name 
-    def populate_coin_list(self):
-        json = self.api_requester.get(
-            'https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest',
-            {'start':'1',
-            'limit':'5000',
-            'convert':'USD'})
-        data = json['data']
-        for coin in data:
-            self.coin_hash_table[coin['name']] = coin['name']
-            self.coin_hash_table[coin['symbol']] = coin['name']
+    def populate_coin_hash(self):
+        try:
+            self.api_requester.open()
+            json = self.api_requester.get(
+                'https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest',
+                {'start':'1',
+                'limit':'5000',
+                'convert':'USD'})
+        except:
+            print("Error in JSON request.")
+        finally:
+            self.api_requester.close()
+            if json != None:
+                data = json['data']
+                for coin in data:
+                    self.coin_hash_table[coin['name']] = coin['name']
+                    self.coin_hash_table[coin['symbol']] = coin['name']
